ensure modifiers are unique
For the `mask-*` utilities, we do call `suggest(…)` multiple times, so
there could be some duplication. The `modifiers` are a simple array and
not a Set so let's make them unique here.

Ideally we don't call suggest with the same values, but that requires a
bigger internal refactor.

Another ideal situation would be if `modifiers` was a Set in the first
place. But since it's export and "public" API I didn't want to change
that datastructure.