Use `default` export condition for `@tailwindcss/vite` (#18948)
## Summary

In `@tailwindcss/vite` 's `package.json`, change the `exports` key from
`include` to `default` since there is no `require` case.

Ran into an issue using the `tsx` package to run a script that has a
sub-dependency that imports from `@tailwindcss/vite`, where `tsx`
converts things to cjs to run, and since there is no `require` case for
this package, it can't find the file. Changing to `default` covers the
cases for both `import` and `require`.

## Test plan

No testing needed. Functionality is the same.

---------

Co-authored-by: Jordan Pittman <jordan@cryptica.me>