use regex replace
The `version` could be _ahead_ when we are preparing a new release. The
version in package.json will point to e.g. 4.1.4 while the latest
version (on npm) is 4.1.3

We _know_ that in these tests we have `v3` as the current package, not
`v3.0.0` so the regex should not change these values.