move parts of `migrate-modernize-arbitrary-values` to core
The part where we upgrade old v3 syntax like `group-[]:` still exists in
the `@tailwindcss/upgrade` package.

We were relying on `postcss-selector-parser`, but the copied part is now
using our own `SelectorParser`.

I also added a temporary attribute selector parser to make everything
pass. Don't look at the code for that _too_ much because we will swap it
out in a later commit 😅