drop unnecessary calls
The the scenario of creating signatures for a component, we always want
to respect the important flag if it is given. Therefore I think that we
can drop the `designSystem.compileAstNodes(candidate,
CompileAstFlags.None)` which saves us some time.

Also removed the `structuredClone` call. Whenever we handle `@apply`
which essentially happened the step before, we already start from a
cloned node internally. This call just meant we were doing it again.