add `cloneAstNode` helper
Our AST nodes are simple plain objects with a `kind` as the discriminant
instead of of an actual class instance. Since we're dealing with simple
plain objects, there is no need for a heavy handed `structuredClone` and
we can use this function instead.

Also added some benchmarks which you can run to see how much the
difference is. On my M1 Pro Max while using battery, these are the
results:

```
  ✓ Cloning AST nodes (2) 2462ms
     name                         hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · cloneAstNode()     5,769,838.55  0.0000  0.0467  0.0002  0.0002  0.0004  0.0004  0.0005  ±0.09%  2884920   fastest
   · structuredClone()    222,598.38  0.0042  4.5958  0.0045  0.0045  0.0048  0.0049  0.0100  ±1.81%   111300

 BENCH  Summary

  cloneAstNode() - ast.bench.ts > Cloning AST nodes
    25.92x faster than structuredClone()
```