wrap color-mix supports declaration in `&`
This is not strictly needed to make it work, but without it the Safari
devtools show weird results.

It is a Safari browser bug, but this is a small enough workaround that
it is worth adding ourselves.

If we ever handle flattening ourselves, then this could in theory also
go away.