normalize percentages when computing the signature
We are only interested in percentages containing a `.`. This way we can
normalize `10.0%` as `10%` by using `Number(value)` and re-attaching the
`%`.

With this, a `10.0%` and `10%` would be considered the same.

The regex in theory could include a leading `[+-]` but that won't have
an effect here, and could be even more confusing when it's being used in
a `calc(…)` expression.