parse `/` as a `word` in ValueParser
Typically each math operator is parsed as a separate word and that's
because it's also typically surrounded by spaces. In case of `/`, this
was parsed as a separator instead of a word including the whitespace (if
any).

The use case we had was to parse `theme(colors.red.500/20%)` where we
don't want `colors.red.500/20%` to be a single word, but we want 3 parts
instead.

This normalizes this a little bit such that all math operators (+, -, *,
/) are just the `word` nodes and any surrounding whitespace will be a
separator node instead.