canonicalize using `SignatureOptions`
Instead of a top-level `DesignSystem`, we can use the
`SignatureOptions`. We can create `CanonicalizeOptions` or similar in
the future that contains `SignatureOptions` _if_ we need values while
canonicalizing values but aren't needed when creating a signature. Just
didn't want to add this as long as we don't need it.

This also changes the APIs for utility and variant canonicalization
steps. The APIs look like this:

```ts
type VariantCanonicalizationFunction = (variant: Variant, options: SignatureOptions) => Variant | Variant[]
type UtilityCanonicalizationFunction = (candidate: Candidate, options: SignatureOptions) => Candidate
```