introduce generalized walk method
This is completely generic, the only requirement is that "parent" nodes
have `{ nodes: T[] }` as the signature.

It's a new iterative walk with minimal memory usage and performance
better than the recursive based walk function we had.

This also adds new functionality where we now can have an `enter` and
`exit` phase. The `exit` phase is like a depth-first where we visit leaf
nodes first.

If you have both, we only do a single walk down and up again instead of
you having to write 2 separate walks that traverse the entire tree.

Some API changes:

1. `replaceWith(…)` doesn't exist anymore, instead you return any of the
   `WalkAction`s. E.g.: `return WalkAction.Replace(…)`
2. `path` is not build while traversing, instead it's a function you can
   call to lazily compute it because in most cases we don't even need
   this functionality.

One benefit is that there is no call stack, so the moment you stop a
walk it's an instant return without unwinding the call stack.

All actions you can do are:

- `WalkAction.Continue`
- `WalkAction.Skip`
- `WalkAction.Stop`
- `WalkAction.Replace(…)`       — replace nodes and continue walking the new nodes
- `WalkAction.ReplaceSkip(…)`   — replace nodes and skip walking the new nodes
- `WalkAction.ReplaceStop(…)`   — replace nodes and stop the entire traversal