do not destructure from `context`
While refactoring, I ran into issues because sometimes we destructure
`parent` which happens to be a "global" value that TypeScript is fine
with if it's undefined (tests would crash).

Using `ctx`, also helped to know where we are using all context related
values.

This will make future commits a bit more readable as well.

Might go back to a destructured pattern, but let's do this for now. In
future commits you will also see that often the `ctx` is just gone,
because instead of using `ctx.replaceWith()` becomes a `return …`