extends `ctx` of CSS AstNode's
The CSS AST has `context` nodes that you can access while walking the
tree, where they build up information over time.

Additionally, when looking at the `path` or the `parent` all the
`context` nodes are hidden.

This adds a small wrapper for the context object when walking over
AstNode's and have to access the `.context`, `.parent` or `.path()`